Major changes and v2.2
All classes were converted to frozen data classes, so that the objects generated by parsing files are somewhat immutable. A very small bug fix occurred in the parsing regex, as a misplaced parenthesis was preventing the parsing of files with missing date fields. Finally, a basic test suite was added.